# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lex Superior - Zimbabwe Civil Law AI Legal Advocate API
 * OpenAPI spec version: 0.1.0
 */
import type { BookmarkType } from "./bookmarkType";

export interface Bookmark {
  id: string;
  userId?: string;
  type: BookmarkType;
  referenceId: string;
  title: string;
  notes?: string;
  createdAt: Date;
}