# lib/api-zod/src/generated/types/createConsultationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lex Superior - Zimbabwe Civil Law AI Legal Advocate API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateConsultationRequest {
  userId?: string;
  title?: string;
  practiceArea: string;
}