# lib/api-zod/src/generated/types/documentGenerationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lex Superior - Zimbabwe Civil Law AI Legal Advocate API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseDetails } from "./caseDetails";

export interface DocumentGenerationRequest {
  documentType: string;
  caseDetails: CaseDetails;
  additionalInfo?: string;
  practiceArea?: string;
}