# lib/api-zod/src/generated/types/statute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lex Superior - Zimbabwe Civil Law AI Legal Advocate API
 * OpenAPI spec version: 0.1.0
 */
import type { StatuteSections } from "./statuteSections";

export interface Statute {
  id: string;
  title: string;
  chapter?: string;
  category: string;
  sections?: StatuteSections;
  summary?: string;
  lastUpdated?: string;
  tags?: string[];
}